#10 Refactor: extract code to add doc to dict.
When dealing with fake stores, we frequently have to add values to a dict using
some specific arguments that make sense to stores (path, name, content etc.) We
did it in at least three places, so now we are extracting this logic to a
specific function. Better yet, we can make it a context manager, leading to
cleaner code.